Change classify.cc so main takes in command-line arguments
Change format of finding accuracy of the model to reading in from the command-line. Add tests
in test-classifier and adjust code according to results of the tests. Move code from classify.cc
into classifier.h function GetClassifierAccuracy to improve the readability of the code. Change
the format of some functions in model.h to make testing in test-classifier easier.