Create file_reader to check validity of files
Change format of finding accuracy of the model to reading in from the command-line. Add tests

Add functions DoesFileExist, CountFileLines, and IsValidImageLabelFile into file_reader to help
determine if the training/classify files fit the specifications. Add test-file-reader.cc to test
functions in file-reader, but noticed that the files in the tests/data folder do not allow empty
spaces, so does not work as expected. However, functions fine in the context of classify.cc.